delete/map_validation(#78)
