+++ inc/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: yshimazu <yshimazu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/24 12:00:27 by user42            #+#    #+#             */
-/*   Updated: 2022/03/14 12:27:58 by yshimazu         ###   ########.fr       */
+/*   Updated: 2022/03/14 18:23:08 by yshimazu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -132,8 +132,8 @@ void	check_args(int ac, char **av);
 void	init_game(t_game *game, char *file_path);
 void	init_image(t_img *img, void *mlx_ptr, int width, int height);
 void	init_design(t_clst *file_lst, int design_end_line, void *mlx_ptr, t_game *game);
-void	init_map(t_clst *file_lst,
-	int map_start_line, int num_nodes, t_game *game);
+void	init_map(t_clst *file_lst, int map_start_line, int num_nodes, t_game *game);
+void	check_map_closed(char **map, t_game *game);
 void	init_minimap(t_minimap *mini, t_map *map, t_point player_pos);
 void	init_player(t_player *player, char **map_ptr, t_game *game);
 void	parse_file(char *file_path, t_game *game, void *mlx_ptr);
